+++ audionimbus/src/bevy/asset/mod.rs
@@ -13,19 +13,19 @@ pub use scene::*;
 /// Constructs a [`SerializedObject`] from an owned byte buffer and passes it to `load`.
 pub(crate) fn with_serialized_object<T>(
     context: &Context,
-    mut bytes: Vec<u8>,
+    bytes: Vec<u8>,
     load: impl FnOnce(&SerializedObject) -> Result<T, SteamAudioError>,
 ) -> Result<T, SteamAudioError> {
-    let serialized_object = SerializedObject::try_with_buffer(context, &mut bytes)?;
+    let serialized_object = SerializedObject::try_with_buffer(context, bytes)?;
     load(&serialized_object)
 }
 
 /// Constructs a [`SerializedObject`] from an owned byte buffer and passes it mutably to `load`.
 pub(crate) fn with_serialized_object_mut<T>(
     context: &Context,
-    mut bytes: Vec<u8>,
+    bytes: Vec<u8>,
     load: impl FnOnce(&mut SerializedObject) -> Result<T, SteamAudioError>,
 ) -> Result<T, SteamAudioError> {
-    let mut serialized_object = SerializedObject::try_with_buffer(context, &mut bytes)?;
+    let mut serialized_object = SerializedObject::try_with_buffer(context, bytes)?;
     load(&mut serialized_object)
 }